chore: split code to be a little easier to maintain later
